refactored: transactions
