+++ src/screens/home/components/transactions/components/desktop/index.tsx
@@ -1,24 +1,60 @@
 import React from 'react';
 import classnames from 'classnames';
+import numeral from 'numeral';
+import dayjs from '@utils/dayjs';
 import useTranslation from 'next-translate/useTranslation';
+import Link from 'next/link';
 import {
+  Typography,
   TableRow,
   TableHead,
   TableCell,
   Table,
   TableBody,
 } from '@material-ui/core';
-import { useTransactionsContext } from '@src/screens/home/components/transactions/contexts/transactions';
+import { getMiddleEllipsis } from '@utils/get_middle_ellipsis';
+import {
+  BLOCK_DETAILS, TRANSACTION_DETAILS,
+} from '@utils/go_to_page';
+import { Result } from '@components';
 import { useStyles } from './styles';
 import { columns } from './utils';
+import { TransactionType } from '../../types';
 
 const Desktop: React.FC<{
   className?: string;
-}> = ({ className }) => {
+  items: TransactionType[];
+}> = ({
+  className, items,
+}) => {
   const classes = useStyles();
-  const { formatUi } = useTransactionsContext();
   const { t } = useTranslation('transactions');
-  const uiData = formatUi('desktop');
+
+  const formattedData = items.map((x) => {
+    return ({
+      block: (
+        <Link href={BLOCK_DETAILS(x.height)} passHref>
+          <Typography variant="body1" component="a">
+            {numeral(x.height).format('0,0')}
+          </Typography>
+        </Link>
+      ),
+      hash: (
+        <Link href={TRANSACTION_DETAILS(x.hash)} passHref>
+          <Typography variant="body1" component="a">
+            {getMiddleEllipsis(x.hash, {
+              beginning: 15, ending: 5,
+            })}
+          </Typography>
+        </Link>
+      ),
+      result: (
+        <Result success={x.success} />
+      ),
+      time: dayjs.utc(x.timestamp).fromNow(),
+      messages: numeral(x.messages).format('0,0'),
+    });
+  });
 
   return (
     <div className={classnames(className, classes.root)}>
@@ -37,7 +73,7 @@ const Desktop: React.FC<{
           </TableRow>
         </TableHead>
         <TableBody>
-          {uiData.map((row, i) => (
+          {formattedData.map((row, i) => (
             <TableRow key={`row-${i}`}>
               {columns.map((column, index) => {
                 const {
