+++ src/screens/home/components/transactions/types.ts
@@ -1,0 +1,11 @@
+export type TransactionType = {
+  height: number;
+  hash: string;
+  success: boolean;
+  timestamp: string;
+  messages: number;
+}
+
+export type TransactionsState = {
+  items: TransactionType[]
+}
